[Refactor]: 타입 및 주석 리팩토링 - 흩어져있던 interface를 interface 디렉토리로 넣어서 타입으로 설정 - 사용하지 않는 import제거 - createUser에서 프론트 루트페이지로 리다이렉션 처리 - 필요 없는 주석 제거
