+++ nestjs/src/auth/interfaces/signInToken.interface.ts
@@ -1,0 +1,4 @@
+export default interface signInToken {
+  token: string;
+  redirectUrl: string;
+}
