+++ nestjs/src/passports/interface/jwtPayload.interface.ts
@@ -1,0 +1,4 @@
+export default interface JwtPayload {
+  id: number;
+  nickname: string;
+}
