+++ nestjs/src/passports/interface/tempJwtPayload.interface.ts
@@ -1,0 +1,3 @@
+export default interface TempJwtPayload {
+  username: string;
+}
